refactor: directories
