+++ server/Recruit-Api/src/main/java/com/econovation/recruit/api/card/service/NavigationService.java
@@ -1,6 +1,6 @@
-package com.econovation.recruit.application.service;
+package com.econovation.recruit.api.card.service;
 
-import com.econovation.recruit.application.port.in.NavigationUseCase;
+import com.econovation.recruit.api.card.usecase.NavigationUseCase;
 import com.econovation.recruitdomain.domains.board.domain.Navigation;
 import com.econovation.recruitdomain.out.NavigationLoadPort;
 import com.econovation.recruitdomain.out.NavigationRecordPort;
