+++ server/Recruit-Api/src/main/java/com/econovation/recruit/api/score/controller/ScoreController.java
@@ -4,7 +4,7 @@ import static com.econovation.recruitcommon.consts.RecruitStatic.SCORE_SUCCESS_R
 import static com.econovation.recruitcommon.consts.RecruitStatic.SCORE_SUCCESS_UPDATE_MESSAGE;
 
 import com.econovation.recruit.api.score.docs.ScoreExceptionDocs;
-import com.econovation.recruit.application.port.in.ScoreUseCase;
+import com.econovation.recruit.api.score.usecase.ScoreUseCase;
 import com.econovation.recruitcommon.annotation.ApiErrorExceptionsExample;
 import com.econovation.recruitdomain.domains.dto.CreateScoreDto;
 import com.econovation.recruitdomain.domains.dto.ScoreAverageDto;
