+++ server/Recruit-Api/src/main/java/com/econovation/recruit/api/score/service/ScoreService.java
@@ -1,8 +1,8 @@
-package com.econovation.recruit.application.service;
+package com.econovation.recruit.api.score.service;
 
 import static com.econovation.recruitcommon.consts.RecruitStatic.CRETERIA_SET;
 
-import com.econovation.recruit.application.port.in.ScoreUseCase;
+import com.econovation.recruit.api.score.usecase.ScoreUseCase;
 import com.econovation.recruit.config.security.SecurityUtils;
 import com.econovation.recruitdomain.domains.dto.CreateScoreDto;
 import com.econovation.recruitdomain.domains.dto.ScoreAverageDto;
